% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faSummary.R
\name{print.faSummary}
\alias{print.faSummary}
\title{Print a factor analytic summary}
\usage{
\method{print}{faSummary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"faSummary"}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a factor analytic summary
}
